reorganized directory structure. added new font for icons
