+++ src/Profitocracy.Mobile/AppInit.xaml.cs
@@ -5,7 +5,7 @@ using Profitocracy.Core.Persistence;
 using Profitocracy.Mobile.Abstractions;
 using Profitocracy.Mobile.Constants;
 using Profitocracy.Mobile.Services;
-using Profitocracy.Mobile.Views.Settings.ProfilesSettings;
+using Profitocracy.Mobile.Views.Settings.Pages;
 
 namespace Profitocracy.Mobile;
