+++ src/Profitocracy.Mobile/Resources/Styles/Styles.xaml
@@ -418,11 +418,13 @@
     <!-- Application-specific styles -->
     
     <!-- Common -->
-    <Style TargetType="Border" x:Key="CloseButton">
+    <Style TargetType="Button" x:Key="CloseButton">
         <Setter Property="WidthRequest" Value="24"/>
         <Setter Property="HeightRequest" Value="24"/>
-        <Setter Property="StrokeThickness" Value="0"/>
-        <Setter Property="StrokeShape" Value="RoundRectangle 16"/>
+        <Setter Property="MinimumWidthRequest" Value="16"/>
+        <Setter Property="MinimumHeightRequest" Value="16"/>
+        <Setter Property="CornerRadius" Value="12"/>
+        <Setter Property="Padding" Value="4"/>
         <Setter Property="BackgroundColor" Value="{StaticResource Gray300}"/>
     </Style>
