--- src/Profitocracy.Mobile/Views/Overview/OverviewPage.xaml
@@ -1,94 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<abstractions:BaseContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:lvc="clr-namespace:LiveChartsCore.SkiaSharpView.Maui;assembly=LiveChartsCore.SkiaSharpView.Maui"
-             xmlns:resx="clr-namespace:Profitocracy.Mobile.Resources.Strings"
-             xmlns:abstractions="clr-namespace:Profitocracy.Mobile.Abstractions"
-             xmlns:viewmodel="clr-namespace:Profitocracy.Mobile.ViewModels.Overview"
-             x:Class="Profitocracy.Mobile.Views.Overview.OverviewPage"
-             x:DataType="viewmodel:OverviewPageViewModel"
-             BackgroundColor="{AppThemeBinding Light={StaticResource LightBackground}, Dark={StaticResource DarkBackground}}"
-             Title="{x:Static resx:AppResources.Pages_Overview}"
-             NavigatedTo="OverviewPage_OnNavigatedTo"
-             HideSoftInputOnTapped="True">
-    <ScrollView VerticalScrollBarVisibility="Never">
-        <StackLayout Padding="8,16">
-            <Picker x:Name="CalculationTypePicker"
-                    SelectedItem="{Binding SelectedDisplayCalculationType}"
-                    ItemDisplayBinding="{Binding DisplayName}"
-                    SelectedIndexChanged="CalculationTypePicker_OnSelectedIndexChanged"/>
-            <Border Style="{StaticResource ExpenseCard}"
-                    Margin="0,16,0,0">
-                <StackLayout>
-                    <Label Text="{x:Static resx:AppResources.Overview_CategoriesExpenses}"
-                           Style="{StaticResource SettingsGroupTitle}"/>
-                    <lvc:CartesianChart Series="{Binding CategoriesExpenses}"
-                                        XAxes="{Binding CategoriesExpensesLabels}"
-                                        HeightRequest="320"
-                                        TooltipPosition="Top"/>
-                </StackLayout>
-            </Border>
-            
-            <Border Style="{StaticResource ExpenseCard}"
-                    Margin="0, 16, 0, 0">
-                <StackLayout>
-                    <Label Text="{x:Static resx:AppResources.Overview_SpendingTypes}"
-                           Style="{StaticResource SettingsGroupTitle}"/>
-                    <lvc:PieChart Series="{Binding SpendingTypeDistribution}"
-                                  HeightRequest="160"
-                                  TooltipPosition="Top"/>
-                </StackLayout>
-            </Border>
-            
-            <Border Style="{StaticResource ExpenseCard}"
-                    Margin="0, 16, 0, 0">
-                <StackLayout>
-                    <Label Text="{x:Static resx:AppResources.Overview_IncomeExpense}"
-                           Style="{StaticResource SettingsGroupTitle}"/>
-                    <lvc:PieChart Series="{Binding IncomeAndExpense}"
-                                  HeightRequest="160"
-                                  TooltipPosition="Top"/>
-                </StackLayout>
-            </Border>
-            
-            <Border Style="{StaticResource ExpenseCard}"
-                    Margin="0, 16, 0, 0">
-                <StackLayout>
-                    <Label Text="{x:Static resx:AppResources.Overview_CategoriesPlannedAndActual}"
-                           Style="{StaticResource SettingsGroupTitle}"/>
-                    <lvc:CartesianChart Series="{Binding PlannedAndActualCategoriesAmounts}"
-                                        XAxes="{Binding PlannedAndActualCategoriesAmountsLabels}"
-                                        HeightRequest="320"
-                                        TooltipPosition="Top"/>
-                </StackLayout>
-            </Border>
-            
-            <Border Style="{StaticResource ExpenseCard}"
-                    Margin="0, 16, 0, 0"
-                    IsVisible="{Binding IsShowDailyExpenses}">
-                <StackLayout>
-                    <Label Text="{x:Static resx:AppResources.Overview_DailyExpenses}"
-                           Style="{StaticResource SettingsGroupTitle}"/>
-                    <lvc:CartesianChart Series="{Binding DailyExpenses}"
-                                        XAxes="{Binding DailyExpensesLabels}"
-                                        HeightRequest="320"
-                                        TooltipPosition="Top"/>
-                </StackLayout>
-            </Border>
-            
-            <Border Style="{StaticResource ExpenseCard}"
-                    Margin="0, 16, 0, 0"
-                    IsVisible="{Binding IsShowWeeklyExpenses}">
-                <StackLayout>
-                    <Label Text="{x:Static resx:AppResources.Overview_WeeklyExpenses}"
-                           Style="{StaticResource SettingsGroupTitle}"/>
-                    <lvc:CartesianChart Series="{Binding WeeklyExpenses}"
-                                        XAxes="{Binding WeeklyExpensesLabels}"
-                                        HeightRequest="320"
-                                        TooltipPosition="Top"/>
-                </StackLayout>
-            </Border>
-        </StackLayout>
-    </ScrollView>
-</abstractions:BaseContentPage>
