--- src/Profitocracy.Mobile/Views/Settings/CategoriesSettings/EditExpenseCategoryPage.xaml
@@ -1,67 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<abstractions:BaseContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-                              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-                              xmlns:ios="clr-namespace:Microsoft.Maui.Controls.PlatformConfiguration.iOSSpecific;assembly=Microsoft.Maui.Controls"
-                              xmlns:resx="clr-namespace:Profitocracy.Mobile.Resources.Strings"
-                              xmlns:abstractions="clr-namespace:Profitocracy.Mobile.Abstractions"
-                              xmlns:viewmodel="clr-namespace:Profitocracy.Mobile.ViewModels.Categories"
-                              x:Class="Profitocracy.Mobile.Views.Settings.CategoriesSettings.EditExpenseCategoryPage"
-                              x:DataType="viewmodel:EditExpenseCategoryPageViewModel"
-                              Padding="16"
-                              Shell.PresentationMode="ModalAnimated"
-                              ios:Page.ModalPresentationStyle="PageSheet"
-                              Loaded="EditExpenseCategoryPage_OnLoaded"
-                              HideSoftInputOnTapped="True">
-    <abstractions:BaseContentPage.Content>
-        <Grid VerticalOptions="Fill"
-              RowDefinitions="auto,*,auto">
-            <FlexLayout Grid.Row="0"
-                        JustifyContent="SpaceBetween">
-                <Label Style="{StaticResource PageTitle}"
-                       Text="{x:Static resx:AppResources.Pages_NewCategory}"/>
-                <Border Style="{StaticResource CloseButton}"
-                        VerticalOptions="End"
-                        x:Name="CloseButton">
-                    <Border.GestureRecognizers>
-                        <TapGestureRecognizer Tapped="CloseButton_OnClicked" />
-                    </Border.GestureRecognizers>
-                    <Label Style="{StaticResource CloseButtonContent}"
-                           Text="X"/>
-                </Border>
-            </FlexLayout>
-            
-            <ScrollView Grid.Row="1">
-                <StackLayout>
-                    <StackLayout Margin="0,32,0,0">
-                        <Label Text="{x:Static resx:AppResources.AddCategory_CategoryName}"/>
-                        <Entry Margin="0,4,0,0"
-                               Text="{Binding Name, Mode=TwoWay}"/>
-                    </StackLayout>
-                
-                    <FlexLayout Margin="0,16,0,0"
-                                JustifyContent="SpaceBetween">
-                        <Label Text="{x:Static resx:AppResources.AddCategory_SpecifyPlannedAmount}" 
-                               VerticalTextAlignment="Center"/>
-                        <Switch IsToggled="{Binding IsPlannedAmountPresent, Mode=TwoWay}"/>    
-                    </FlexLayout>
-                
-                
-                    <StackLayout Margin="0,16,0,0"
-                                 IsVisible="{Binding IsPlannedAmountPresent}">
-                        <Label Text="{x:Static resx:AppResources.AddCategory_PlannedAmount}"/>
-                        <Entry Margin="0,4,0,0"
-                               Keyboard="Numeric"
-                               Text="{Binding PlannedAmount, Mode=TwoWay}"/>    
-                    </StackLayout>
-                </StackLayout>
-            </ScrollView>
-            
-            <Button Grid.Row="2"
-                    Text="{x:Static resx:AppResources.AddCategory_SaveCategory}"
-                    x:Name="AddCategoryButton"
-                    VerticalOptions="End"
-                    Clicked="AddCategoryButton_OnClicked"/>
-        </Grid>
-    </abstractions:BaseContentPage.Content>
-</abstractions:BaseContentPage>
