--- src/Profitocracy.Mobile/Views/Settings/CategoriesSettings/ExpenseCategoriesSettingsPage.xaml
@@ -1,107 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<abstractions:BaseContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-                              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-                              xmlns:resx="clr-namespace:Profitocracy.Mobile.Resources.Strings"
-                              xmlns:categories="clr-namespace:Profitocracy.Mobile.Models.Categories"
-                              xmlns:abstractions="clr-namespace:Profitocracy.Mobile.Abstractions"
-                              xmlns:viewmodel="clr-namespace:Profitocracy.Mobile.ViewModels.Categories"
-                              x:Class="Profitocracy.Mobile.Views.Settings.CategoriesSettings.ExpenseCategoriesSettingsPage"
-                              x:DataType="viewmodel:ExpenseCategoriesSettingsPageViewModel"
-                              Title="{x:Static resx:AppResources.Pages_Categories}"
-                              BackgroundColor="{AppThemeBinding Light={StaticResource LightBackground}, Dark={StaticResource DarkBackground}}"
-                              NavigatedTo="UpdateCategoriesList"
-                              Padding="8,0,8,0"
-                              HideSoftInputOnTapped="True">
-    <abstractions:BaseContentPage.Content>
-        <Grid>
-            <Grid.RowDefinitions>
-                <RowDefinition Height="*"/>
-            </Grid.RowDefinitions>
-            
-            <CollectionView Grid.Row="0"
-                            x:Name="CategoriesCollectionView"
-                            VerticalScrollBarVisibility="Never">
-                <CollectionView.ItemTemplate>
-                    <DataTemplate x:DataType="categories:CategoryModel">
-                        <SwipeView>
-                            <SwipeView.RightItems>
-                                <SwipeItems>
-                                    <SwipeItemView Padding="4,0,8,0"
-                                                   BackgroundColor="Transparent"
-                                                   Invoked="DeleteCategory_OnInvoked">
-                                        <VerticalStackLayout VerticalOptions="Center" 
-                                                             Width="16">
-                                            <Border BackgroundColor="{StaticResource DangerRed}"
-                                                    StrokeThickness="0"
-                                                    Padding="16">
-                                                <Border.StrokeShape>
-                                                    <RoundRectangle CornerRadius="40"/>
-                                                </Border.StrokeShape>
-                                                <Image Source="bin.png" 
-                                                       Width="8"/>
-                                            </Border>
-                                        </VerticalStackLayout>
-                                    </SwipeItemView>
-                                    <SwipeItemView Padding="8,0,4,0"
-                                                   BackgroundColor="Transparent"
-                                                   Invoked="EditCategory_OnInvoked">
-                                        <VerticalStackLayout VerticalOptions="Center" 
-                                                             Width="16">
-                                            <Border BackgroundColor="{StaticResource InfoBlue}"
-                                                    StrokeThickness="0"
-                                                    Padding="16">
-                                                <Border.StrokeShape>
-                                                    <RoundRectangle CornerRadius="40"/>
-                                                </Border.StrokeShape>
-                                                <Image Source="edit.png" 
-                                                       Width="8"/>
-                                            </Border>
-                                        </VerticalStackLayout>
-                                    </SwipeItemView>
-                                </SwipeItems>
-                            </SwipeView.RightItems>
-                            
-                            <Border Style="{StaticResource CategoryCard}"
-                                    Margin="0,8,0,0">
-                                <StackLayout>
-                                    <FlexLayout JustifyContent="SpaceBetween">
-                                        <Label Style="{StaticResource CategoryCardName}"
-                                               Text="{Binding Name}"/>
-                                        <Label Style="{StaticResource CategoryCardPlannedAmount}"
-                                               Text="{Binding DisplayPlannedAmount}"/>
-                                    </FlexLayout>
-                                </StackLayout>
-                            </Border>
-                        </SwipeView>
-                    </DataTemplate>
-                </CollectionView.ItemTemplate>
-                <CollectionView.Footer>
-                    <Grid HeightRequest="92"
-                          BackgroundColor="Transparent"/>
-                </CollectionView.Footer>
-            </CollectionView>
-            
-            <Button Grid.Row="0"
-                    Text="+"
-                    HorizontalOptions="End"
-                    VerticalOptions="End"
-                    FontAttributes="Bold"
-                    BackgroundColor="{StaticResource Primary}"
-                    WidthRequest="60"
-                    HeightRequest="60"
-                    CornerRadius="30"
-                    Margin="0,0,8,16"
-                    FontSize="25"
-                    x:Name="AddCategoryButton"
-                    Clicked="AddCategoryButton_OnClicked">
-                <Button.Shadow>
-                    <Shadow Brush="{AppThemeBinding Light={StaticResource Gray300}, Dark={StaticResource Gray950}}"
-                            Offset="0,8"
-                            Radius="16"
-                            Opacity="0.9" />
-                </Button.Shadow>
-            </Button>
-        </Grid>
-    </abstractions:BaseContentPage.Content>
-</abstractions:BaseContentPage>
