--- src/Profitocracy.Mobile/Views/Settings/LanguageSettings/LanguageSettingsPage.xaml
@@ -1,94 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<abstractions:BaseContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-                              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-                              xmlns:resx="clr-namespace:Profitocracy.Mobile.Resources.Strings"
-                              xmlns:abstractions="clr-namespace:Profitocracy.Mobile.Abstractions"
-                              xmlns:viewmodel="clr-namespace:Profitocracy.Mobile.ViewModels.Settings"
-                              x:Class="Profitocracy.Mobile.Views.Settings.LanguageSettings.LanguageSettingsPage"
-                              x:DataType="viewmodel:LanguageSettingsViewModel"
-                              Title="{x:Static resx:AppResources.Pages_Language}"
-                              Loaded="LanguageSettingsPage_OnLoaded"
-                              BackgroundColor="{AppThemeBinding Light={StaticResource LightBackground}, Dark={StaticResource DarkBackground}}"
-                              HideSoftInputOnTapped="True">
-    <abstractions:BaseContentPage.Content>
-        <ScrollView Padding="16"
-                VerticalScrollBarVisibility="Never">
-            <StackLayout>
-                <Border Style="{StaticResource ThemeSettingsButton}"
-                        Margin="0,8,0,0" >
-                    <Border.GestureRecognizers>
-                        <TapGestureRecognizer Tapped="EnglishLanguage_OnSelected"/>
-                    </Border.GestureRecognizers>
-                    <Border.Triggers>
-                        <DataTrigger TargetType="Border"
-                                     Binding="{Binding IsEnglish}"
-                                     Value="True">
-                            <Setter Property="Style" Value="{StaticResource ThemeSettingsButtonActive}"/>
-                        </DataTrigger>
-                        <DataTrigger TargetType="Border"
-                                     Binding="{Binding IsEnglish}"
-                                     Value="False">
-                            <Setter Property="Style" Value="{StaticResource ThemeSettingsButton}"/>
-                        </DataTrigger>
-                    </Border.Triggers>
-                    <FlexLayout JustifyContent="SpaceBetween">
-                        <Label Style="{StaticResource SettingsButtonText}"
-                               Text="{x:Static resx:AppResources.Languages_English}" 
-                               VerticalOptions="Center">
-                            <Label.Triggers>
-                                <DataTrigger TargetType="Label"
-                                             Binding="{Binding IsEnglish}"
-                                             Value="True">
-                                    <Setter Property="TextColor" Value="{StaticResource DarkPrimaryText}"/>
-                                </DataTrigger>
-                                <DataTrigger TargetType="Label"
-                                             Binding="{Binding IsEnglish}"
-                                             Value="False">
-                                    <Setter Property="TextColor" Value="{AppThemeBinding Light={StaticResource LightPrimaryText}, Dark={StaticResource DarkPrimaryText}}"/>
-                                </DataTrigger>
-                            </Label.Triggers>   
-                        </Label>
-                    </FlexLayout>
-                </Border>
-                
-                <Border Style="{StaticResource ThemeSettingsButton}"
-                        Margin="0,8,0,0">
-                    <Border.GestureRecognizers>
-                        <TapGestureRecognizer Tapped="RussianLanguage_OnSelected"/>
-                    </Border.GestureRecognizers>
-                    <Border.Triggers>
-                        <DataTrigger TargetType="Border"
-                                     Binding="{Binding IsRussian}"
-                                     Value="True">
-                            <Setter Property="Style" Value="{StaticResource ThemeSettingsButtonActive}"/>
-                        </DataTrigger>
-                        <DataTrigger TargetType="Border"
-                                     Binding="{Binding IsRussian}"
-                                     Value="False">
-                            <Setter Property="Style" Value="{StaticResource ThemeSettingsButton}"/>
-                        </DataTrigger>
-                    </Border.Triggers>
-                    <FlexLayout JustifyContent="SpaceBetween">
-                        <Label Style="{StaticResource SettingsButtonText}"
-                               Text="{x:Static resx:AppResources.Languages_Russian}" 
-                               VerticalOptions="Center">
-                            <Label.Triggers>
-                                <DataTrigger TargetType="Label"
-                                             Binding="{Binding IsRussian}"
-                                             Value="True">
-                                    <Setter Property="TextColor" Value="{StaticResource DarkPrimaryText}"/>
-                                </DataTrigger>
-                                <DataTrigger TargetType="Label"
-                                             Binding="{Binding IsRussian}"
-                                             Value="False">
-                                    <Setter Property="TextColor" Value="{AppThemeBinding Light={StaticResource LightPrimaryText}, Dark={StaticResource DarkPrimaryText}}"/>
-                                </DataTrigger>
-                            </Label.Triggers>   
-                        </Label>
-                    </FlexLayout>
-                </Border>
-            </StackLayout>
-        </ScrollView>
-    </abstractions:BaseContentPage.Content>
-</abstractions:BaseContentPage>
