--- src/Profitocracy.Mobile/Views/Settings/ProfilesSettings/EditProfilePage.xaml
@@ -1,75 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<abstractions:BaseContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-                              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-                              xmlns:ios="clr-namespace:Microsoft.Maui.Controls.PlatformConfiguration.iOSSpecific;assembly=Microsoft.Maui.Controls"
-                              xmlns:resx="clr-namespace:Profitocracy.Mobile.Resources.Strings"
-                              xmlns:abstractions="clr-namespace:Profitocracy.Mobile.Abstractions"
-                              xmlns:viewmodel="clr-namespace:Profitocracy.Mobile.ViewModels.Profiles"
-                              x:Class="Profitocracy.Mobile.Views.Settings.ProfilesSettings.EditProfilePage"
-                              x:DataType="viewmodel:EditProfilePageViewModel"
-                              ios:Page.ModalPresentationStyle="FullScreen"
-                              Shell.PresentationMode="ModalAnimated"
-                              Loaded="EditProfilePage_OnLoaded"
-                              Padding="16"
-                              HideSoftInputOnTapped="True">
-    <Shell.BackButtonBehavior>
-        <BackButtonBehavior IsVisible="False" 
-                            IsEnabled="False" />
-    </Shell.BackButtonBehavior>
-    <abstractions:BaseContentPage.Content>
-        <Grid VerticalOptions="Fill"
-              RowDefinitions="auto,*,auto">
-            <FlexLayout Grid.Row="0"
-                        JustifyContent="SpaceBetween">
-                <Label Style="{StaticResource PageTitle}"
-                       Text="{x:Static resx:AppResources.EditProfile_Profile}"/>
-                <Border Style="{StaticResource CloseButton}"
-                        VerticalOptions="End"
-                        x:Name="CloseButton"
-                        IsEnabled="{Binding IsNotFirstProfile}"
-                        IsVisible="{Binding IsNotFirstProfile}">
-                    <Border.GestureRecognizers>
-                        <TapGestureRecognizer Tapped="CloseButton_OnClicked" />
-                    </Border.GestureRecognizers>
-                    <Label Style="{StaticResource CloseButtonContent}"
-                           Text="X"/>
-                </Border>
-            </FlexLayout>
-            <ScrollView Grid.Row="1"
-                        Margin="0,32,0,0">
-                <StackLayout>
-                    <StackLayout>
-                        <Label Text="{x:Static resx:AppResources.EditProfile_ProfileName}"/>
-                        <Entry Margin="0,4,0,0"
-                               Text="{Binding Name, Mode=TwoWay}" />
-                    </StackLayout>
-            
-                    <StackLayout Margin="0,16,0,0">
-                        <Label Text="{x:Static resx:AppResources.EditProfile_InitialBalance}"/>
-                        <Grid Margin="0,4,0,0">
-                            <Grid.ColumnDefinitions>
-                                <ColumnDefinition />
-                                <ColumnDefinition Width="8"/>
-                                <ColumnDefinition Width="3*" />
-                            </Grid.ColumnDefinitions>
-                                        
-                            <Picker Grid.Column="0"
-                                    x:Name="ProfileCurrencyPicker"
-                                    SelectedItem="{Binding SelectedCurrency}"
-                                    ItemDisplayBinding="{Binding Code}"/>
-                            <Entry Grid.Column="2" Margin="0,4,0,0"
-                                   Keyboard="Numeric"
-                                   Text="{Binding InitialBalance, Mode=TwoWay}"/>
-                        </Grid>
-                    </StackLayout>
-                </StackLayout>
-            </ScrollView>
-            
-            <Button Grid.Row="2"
-                    VerticalOptions="End"
-                    Text="{x:Static resx:AppResources.EditProfile_SaveProfile}" 
-                    Clicked="Button_OnClicked"/>
-        </Grid>
-    </abstractions:BaseContentPage.Content>
-</abstractions:BaseContentPage>
