--- src/Profitocracy.Mobile/Views/Settings/ProfilesSettings/ProfilesSettingsPage.xaml
@@ -1,125 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<abstractions:BaseContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-                              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-                              xmlns:resx="clr-namespace:Profitocracy.Mobile.Resources.Strings"
-                              xmlns:abstractions="clr-namespace:Profitocracy.Mobile.Abstractions"
-                              xmlns:viewmodel="clr-namespace:Profitocracy.Mobile.ViewModels.Profiles"
-                              xmlns:profiles="clr-namespace:Profitocracy.Mobile.Models.Profiles"
-                              x:Class="Profitocracy.Mobile.Views.Settings.ProfilesSettings.ProfilesSettingsPage"
-                              x:DataType="viewmodel:ProfileSettingsPageViewModel"
-                              Title="{x:Static resx:AppResources.Pages_Profiles}"
-                              BackgroundColor="{AppThemeBinding Light={StaticResource LightBackground}, Dark={StaticResource DarkBackground}}"
-                              NavigatedTo="UpdateProfilesList"
-                              Padding="8,0,8,0"
-                              HideSoftInputOnTapped="True">
-    <abstractions:BaseContentPage.Content>
-        <Grid>
-            <Grid.RowDefinitions>
-                <RowDefinition Height="*"/>
-            </Grid.RowDefinitions>
-            
-            <CollectionView Grid.Row="0"
-                            x:Name="ProfilesCollectionView"
-                            VerticalScrollBarVisibility="Never">
-                <CollectionView.ItemTemplate>
-                    <DataTemplate x:DataType="profiles:ProfileModel">
-                        <SwipeView>
-                            <SwipeView.RightItems>
-                                <SwipeItems>
-                                    <SwipeItemView Padding="4,0,8,0"
-                                                   BackgroundColor="Transparent"
-                                                   Invoked="DeleteProfile_OnInvoked">
-                                        <VerticalStackLayout VerticalOptions="Center" 
-                                                             Width="16">
-                                            <Border BackgroundColor="{StaticResource DangerRed}"
-                                                    StrokeThickness="0"
-                                                    Padding="16">
-                                                <Border.StrokeShape>
-                                                    <RoundRectangle CornerRadius="40"/>
-                                                </Border.StrokeShape>
-                                                <Image Source="bin.png" 
-                                                       Width="8"/>
-                                            </Border>
-                                        </VerticalStackLayout>
-                                    </SwipeItemView>
-                                    <SwipeItemView Padding="8,0,4,0"
-                                                   BackgroundColor="Transparent"
-                                                   Invoked="EditProfile_OnInvoked">
-                                        <VerticalStackLayout VerticalOptions="Center" 
-                                                             Width="16">
-                                            <Border BackgroundColor="{StaticResource InfoBlue}"
-                                                    StrokeThickness="0"
-                                                    Padding="16">
-                                                <Border.StrokeShape>
-                                                    <RoundRectangle CornerRadius="40"/>
-                                                </Border.StrokeShape>
-                                                <Image Source="edit.png" 
-                                                       Width="8"/>
-                                            </Border>
-                                        </VerticalStackLayout>
-                                    </SwipeItemView>
-                                </SwipeItems>
-                            </SwipeView.RightItems>
-                            
-                            <StackLayout>
-                                <Border Style="{StaticResource CurrentProfileCard}"
-                                        Margin="0,8,0,0"
-                                        IsVisible="{Binding IsCurrent}">
-                                    <StackLayout>
-                                        <FlexLayout JustifyContent="SpaceBetween">
-                                            <Label Style="{StaticResource CurrentProfileCardName}"
-                                                   Text="{Binding Name}"/>
-                                            <Label Style="{StaticResource CurrentProfileCardName}"
-                                                   Text="{Binding CurrencySymbol}"/>
-                                        </FlexLayout>
-                                    </StackLayout>
-                                </Border> 
-                                <Border Style="{StaticResource ProfileCard}"
-                                        Margin="0,8,0,0"
-                                        IsVisible="{Binding IsNotCurrent}">
-                                    <Border.GestureRecognizers>
-                                        <TapGestureRecognizer Tapped="ProfileCard_OnTapped"/>
-                                    </Border.GestureRecognizers>
-                                    <StackLayout>
-                                        <FlexLayout JustifyContent="SpaceBetween">
-                                            <Label Style="{StaticResource CategoryCardName}"
-                                                   Text="{Binding Name}"/>
-                                            <Label Style="{StaticResource CategoryCardPlannedAmount}"
-                                                   Text="{Binding CurrencySymbol}"/>
-                                        </FlexLayout>
-                                    </StackLayout>
-                                </Border>
-                            </StackLayout>
-                        </SwipeView>
-                    </DataTemplate>
-                </CollectionView.ItemTemplate>
-                <CollectionView.Footer>
-                    <Grid HeightRequest="92" 
-                          BackgroundColor="Transparent"/>
-                </CollectionView.Footer>
-            </CollectionView>
-            
-            <Button Grid.Row="0"
-                    Text="+"
-                    HorizontalOptions="End"
-                    VerticalOptions="End"
-                    FontAttributes="Bold"
-                    BackgroundColor="{StaticResource Primary}"
-                    WidthRequest="60"
-                    HeightRequest="60"
-                    CornerRadius="30"
-                    Margin="0,0,8,16"
-                    FontSize="25"
-                    x:Name="AddProfileButton"
-                    Clicked="AddProfileButton_OnClicked">
-                <Button.Shadow>
-                    <Shadow Brush="{AppThemeBinding Light={StaticResource Gray300}, Dark={StaticResource Gray950}}"
-                            Offset="0,8"
-                            Radius="16"
-                            Opacity="0.9" />
-                </Button.Shadow>
-            </Button>
-        </Grid>
-    </abstractions:BaseContentPage.Content>
-</abstractions:BaseContentPage>
