--- src/Profitocracy.Mobile/Views/Settings/SettingsPage.xaml
@@ -1,96 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<abstractions:BaseContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-                              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-                              xmlns:resx="clr-namespace:Profitocracy.Mobile.Resources.Strings"
-                              xmlns:abstractions="clr-namespace:Profitocracy.Mobile.Abstractions"
-                              x:Class="Profitocracy.Mobile.Views.Settings.SettingsPage"
-                              Title="{x:Static resx:AppResources.Pages_Settings}"
-                              BackgroundColor="{AppThemeBinding Light={StaticResource LightBackground}, Dark={StaticResource DarkBackground}}"
-                              HideSoftInputOnTapped="True">
-    <ScrollView Padding="16"
-                VerticalScrollBarVisibility="Never">
-        <StackLayout>
-            <Label Text="{x:Static resx:AppResources.Settings_Application}"
-                   Style="{StaticResource SettingsGroupTitle}"
-                   Margin="0,0,0,0"/>
-            <Border Style="{StaticResource SettingsButton}"
-                    Margin="0,8,0,0">
-                <Border.GestureRecognizers>
-                    <TapGestureRecognizer Tapped="ProfilesButton_OnClicked"/>
-                </Border.GestureRecognizers>
-                <FlexLayout JustifyContent="SpaceBetween">
-                    <Label Style="{StaticResource SettingsButtonText}"
-                           Text="{x:Static resx:AppResources.Settings_Profiles}" />
-                    <Label Style="{StaticResource SettingsButtonText}"
-                           Text=">" />
-                </FlexLayout>
-            </Border>
-            <Border Style="{StaticResource SettingsButton}"
-                    Margin="0,8,0,0" >
-                <Border.GestureRecognizers>
-                    <TapGestureRecognizer Tapped="CategoriesButton_OnClicked"/>
-                </Border.GestureRecognizers>
-                <FlexLayout JustifyContent="SpaceBetween">
-                    <Label Style="{StaticResource SettingsButtonText}"
-                           Text="{x:Static resx:AppResources.Settings_Categories}" />
-                    <Label Style="{StaticResource SettingsButtonText}"
-                           Text=">" />
-                </FlexLayout>
-            </Border>
-            
-            <Label Text="{x:Static resx:AppResources.Settings_General}"
-                   Style="{StaticResource SettingsGroupTitle}"
-                   Margin="0,16,0,0"/>
-            <Border Style="{StaticResource SettingsButton}"
-                    Margin="0,8,0,0" >
-                <Border.GestureRecognizers>
-                    <TapGestureRecognizer Tapped="ThemeButton_OnClicked"/>
-                </Border.GestureRecognizers>
-                <FlexLayout JustifyContent="SpaceBetween">
-                    <Label Style="{StaticResource SettingsButtonText}"
-                           Text="{x:Static resx:AppResources.Settings_Theme}" />
-                    <Label Style="{StaticResource SettingsButtonText}"
-                           Text=">" />
-                </FlexLayout>
-            </Border>
-            <Border Style="{StaticResource SettingsButton}"
-                    Margin="0,8,0,0" >
-                <Border.GestureRecognizers>
-                    <TapGestureRecognizer Tapped="LanguageButton_OnClicked"/>
-                </Border.GestureRecognizers>
-                <FlexLayout JustifyContent="SpaceBetween">
-                    <Label Style="{StaticResource SettingsButtonText}"
-                           Text="{x:Static resx:AppResources.Settings_Language}" />
-                    <Label Style="{StaticResource SettingsButtonText}"
-                           Text=">" />
-                </FlexLayout>
-            </Border>
-            
-            <Label Text="{x:Static resx:AppResources.Settings_AboutApplication}"
-                   Style="{StaticResource SettingsGroupTitle}"
-                   Margin="0,16,0,0"/>
-            <Border Style="{StaticResource SettingsButton}"
-                    Margin="0,8,0,0" >
-                <Border.GestureRecognizers>
-                    <TapGestureRecognizer Tapped="GitHubButton_OnClicked"/>
-                </Border.GestureRecognizers>
-                <FlexLayout JustifyContent="SpaceBetween">
-                    <Label Style="{StaticResource SettingsButtonText}"
-                           Text="{x:Static resx:AppResources.Settings_GitHub}" />
-                    <Label Style="{StaticResource SettingsButtonText}"
-                           Text=">" />
-                </FlexLayout>
-            </Border>
-            <Border Style="{StaticResource SettingsButton}"
-                    Margin="0,8,0,0" >
-                <FlexLayout JustifyContent="SpaceBetween">
-                    <Label Style="{StaticResource SettingsButtonText}"
-                           Text="{x:Static resx:AppResources.Settings_Version}" />
-                    <Label Style="{StaticResource DisabledSettingsButtonText}"
-                           x:Name="VersionLabel"/>
-                </FlexLayout>
-            </Border>
-        </StackLayout>
-    </ScrollView>
-</abstractions:BaseContentPage>
