+++ src/Profitocracy.Mobile/Views/Shared/Controls/AddNewButton.xaml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<ContentView xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             x:Class="Profitocracy.Mobile.Views.Shared.Controls.AddNewButton"
+             WidthRequest="60"
+             HeightRequest="60">
+    <Button BackgroundColor="{StaticResource Primary}"
+            WidthRequest="60"
+            HeightRequest="60"
+            CornerRadius="30"
+            ImageSource="{FontImage &#xf217;,
+                FontFamily=ionicons,
+                Size=24}"
+            Clicked="InternalAddNewButton_OnClicked">
+        <Button.Shadow>
+            <Shadow Brush="{AppThemeBinding Light={StaticResource Gray300}, Dark={StaticResource Gray950}}"
+                    Offset="0,8"
+                    Radius="16"
+                    Opacity="0.9" />
+        </Button.Shadow>
+    </Button>
+</ContentView>
