+++ src/Profitocracy.Mobile/Views/Shared/Controls/ModalHeaderView.xaml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<ContentView xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             xmlns:controls="clr-namespace:Profitocracy.Mobile.Views.Shared.Controls"
+             x:Class="Profitocracy.Mobile.Views.Shared.Controls.ModalHeaderView"
+             x:DataType="controls:ModalHeaderView">
+    <FlexLayout JustifyContent="SpaceBetween">
+        <Label Style="{StaticResource PageTitle}"
+               Text="{Binding Title}"
+               VerticalTextAlignment="Center"/>
+        
+        <Button Style="{StaticResource CloseButton}"
+                ImageSource="{FontImage &#xf129;,
+                    FontFamily=ionicons,
+                    Size=12}"
+                Clicked="CloseButton_OnClicked"
+                VerticalOptions="Center"/>
+    </FlexLayout>
+</ContentView>
