+++ src/Profitocracy.Mobile/Views/Transactions/Pages/FilteredTransactionsPage.xaml
@@ -1,0 +1,39 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<abstractions:BaseContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+                              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+                              xmlns:ios="clr-namespace:Microsoft.Maui.Controls.PlatformConfiguration.iOSSpecific;assembly=Microsoft.Maui.Controls"
+                              xmlns:resx="clr-namespace:Profitocracy.Mobile.Resources.Strings"
+                              xmlns:transactions="clr-namespace:Profitocracy.Mobile.Models.Transactions"
+                              xmlns:abstractions="clr-namespace:Profitocracy.Mobile.Abstractions"
+                              xmlns:viewmodel="clr-namespace:Profitocracy.Mobile.ViewModels.Transactions"
+                              xmlns:controls="clr-namespace:Profitocracy.Mobile.Views.Transactions.Controls"
+                              xmlns:sharedcontrols="clr-namespace:Profitocracy.Mobile.Views.Shared.Controls"
+                              x:Class="Profitocracy.Mobile.Views.Transactions.Pages.FilteredTransactionsPage"
+                              x:DataType="viewmodel:FilteredTransactionsPageViewModel"
+                              Shell.PresentationMode="ModalAnimated"
+                              ios:Page.ModalPresentationStyle="PageSheet"
+                              HideSoftInputOnTapped="True">
+    <abstractions:BaseContentPage.Content>
+        <Grid>
+            <Grid.RowDefinitions>
+                <RowDefinition Height="auto"/>
+                <RowDefinition Height="*"/>
+            </Grid.RowDefinitions>
+            
+            <StackLayout Grid.Row="0"
+                         Padding="16">
+                <sharedcontrols:ModalHeaderView Title="{x:Static resx:AppResources.Pages_Transactions}" 
+                                                CloseClicked="CloseButton_OnClicked"/>
+            </StackLayout>
+            
+            <CollectionView Grid.Row="1" x:Name="TransactionsCollectionView">
+                <CollectionView.ItemTemplate>
+                    <DataTemplate x:DataType="transactions:TransactionModel">
+                        <controls:TransactionView Transaction="{Binding}"/>
+                    </DataTemplate>
+                </CollectionView.ItemTemplate>
+            </CollectionView>
+        </Grid>
+    </abstractions:BaseContentPage.Content>
+</abstractions:BaseContentPage>
